Add rustdoc comments for functions
