+++ src/utils/sort_by_priority.rs
@@ -1,4 +1,5 @@
 pub trait SortByPriority<'a, U> {
+    /// Sorts the items in the collection by the priority function provided.
     fn sort_by_priority<F>(&mut self, priority_func: F)
     where
         F: Fn(&U) -> u32;
@@ -8,6 +9,7 @@ impl<'a, T, U> SortByPriority<'a, U> for T
 where
     T: AsMut<[U]>,
 {
+    /// Sorts the items in the collection by the priority function provided.
     fn sort_by_priority<F>(&mut self, priority_func: F)
     where
         F: Fn(&U) -> u32,
